Make some new files to separate code
